% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apex_domain}
\alias{apex_domain}
\title{Return the apex/top-private domain from a vector of domains}
\usage{
apex_domain(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
character vector
}
\description{
Return the apex/top-private domain from a vector of domains
}
