% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_public_suffix}
\alias{is_public_suffix}
\title{Test whether a domain is a public suffix}
\usage{
is_public_suffix(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
character vector
}
\description{
Test whether a domain is a public suffix
}
