% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl-package.R
\docType{package}
\name{psl}
\alias{psl}
\alias{psl-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/psl}
\item BugReports: \url{https://gitlab.com/hrbrmstr/psl/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
