% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{public_suffix}
\alias{public_suffix}
\title{Return the public suffix from a vector of domains}
\usage{
public_suffix(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
character vector
}
\description{
Return the public suffix from a vector of domains
}
