% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{suffix_extract}
\alias{suffix_extract}
\title{Separate a domain into component parts}
\usage{
suffix_extract(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
data frame
}
\description{
Separate a domain into component parts
}
