% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{suffix_extract2}
\alias{suffix_extract2}
\title{Separate a domain into component parts}
\usage{
suffix_extract2(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
data frame
}
\description{
Compatibility function for those using \code{urltools::suffix_extract()}
}
